task file and linting
