+++ internal/httpserve/route/static.go
@@ -49,8 +49,6 @@ func (t *Template) Render(w io.Writer, name string, data interface{}, c echo.Con
 //go:embed openapi.json
 //go:embed robots.txt
 //go:embed security.txt
-//go:embed rest.yaml
-//go:embed redoc-static.html
 var openapi embed.FS
 
 // registerOpenAPISpecHandler embeds our generated open api specs and serves it behind /api-docs
@@ -58,28 +56,7 @@ func registerOpenAPISpecHandler(router *echo.Echo) (err error) {
 	_, err = router.AddRoute(echo.Route{
 		Method:  http.MethodGet,
 		Path:    "/api-docs",
-		Handler: echo.StaticFileHandler("openapi.json", openapi),
-	}.ForGroup(V1Version, mw))
-
-	return
-}
-
-func registerRedocHandler(router *echo.Echo) (err error) {
-	_, err = router.AddRoute(echo.Route{
-		Method:  http.MethodGet,
-		Path:    "/redoc",
-		Handler: echo.StaticFileHandler("redoc-static.html", openapi),
-	}.ForGroup(V1Version, mw))
-
-	return
-}
-
-// registerSwaggerStatic embeds our generated open api specs and serves it behind /rest.yaml
-func registerSwaggerStatic(router *echo.Echo) (err error) {
-	_, err = router.AddRoute(echo.Route{
-		Method:  http.MethodGet,
-		Path:    "/restapi-docs",
-		Handler: echo.StaticFileHandler("rest.yaml", openapi),
+		Handler: echo.StaticFileHandler("joined.yaml", openapi),
 	}.ForGroup(V1Version, mw))
 
 	return
