+++ pkg/utils/oas/Taskfile.yaml
@@ -1,0 +1,11 @@
+version: '3'
+
+tasks:
+  join:
+    desc: joins the graph generated OpenAPI specifications and the REST specifications into one file
+    cmds:
+      - redocly join pkg/utils/oas/specifications/rest.yaml pkg/utils/oas/specifications/openapi.json -o joined.yaml --without-x-tag-groups
+  lint:
+    desc: lint OpenAPI specifications
+    cmds: 
+      - vacuum lint ./specifications/openapi.json
