+++ pkg/utils/oas/redoc/README.md
@@ -1,0 +1,7 @@
+# Redocly
+
+Install via `npm i -g @redocly/cli@latest`
+
+Joining multiple OAS files via `redocly join rest.yaml openapi.json -o joined.yaml --without-x-tag-groups`
+
+Build / output documentation via `redocly build-docs joined.yaml  --config="redocly.yaml"`
